Made first big changes from Pull Request revisions.
First thing done was calling Utils::microseconds() once and
storing it in a variable. Next, was getting rid of the
parameter passed into write_data, and instead use
a static variable to hold the times. Finally
I added the stagger times into defaultConfig.txt
and will update them in the TCPManager in the
next commit!