Finished all revisions except for last revision
The revisions of the commit are: implementing the
defaultconfig.txt values into write_data, changing the
name of times to last_sent_times for checking when to
send data. The first revision was done by adding a
ConfiguratorManager check and write to the array
"stagger_times" inside the beginning of tcp_loop.